refactor: split code for category component
